fix(react): fix jsdom react testing issues Fixes vmware#5985 - Add isNode check to property syncing in cdscontrol.
Fixes vmware#6525:
- Remove isNode check for aria-reflect, now that all roles are set in connectedCallback
