+++ packages/core/src/internal/utils/environment.spec.ts
@@ -1,0 +1,9 @@
+import { isBrowser } from './environment';
+
+describe('Environment Helper: ', () => {
+  describe('isBrowser():', () => {
+    it('returns true when expected', () => {
+      expect(isBrowser()).toBe(true);
+    });
+  });
+});
