+++ packages/core/src/internal/utils/environment.ts
@@ -1,0 +1,9 @@
+import { isNil } from '../utils/identity.js';
+
+export function isBrowser(win = window) {
+  return !isNil(win);
+}
+
+export function isJestTest() {
+  return (globalThis as any)?.process?.env?.JEST_WORKER_ID !== undefined;
+}
