+++ packages/core/src/internal/utils/events.ts
@@ -1,8 +1,10 @@
 /*
- * Copyright (c) 2016-2021 VMware, Inc. All Rights Reserved.
+ * Copyright (c) 2016-2022 VMware, Inc. All Rights Reserved.
  * This software is released under MIT license.
  * The full license information can be found in LICENSE in the root directory of this project.
  */
+
+import { isJestTest } from './environment.js';
 
 export function stopEvent(event: Event) {
   event.preventDefault();
@@ -29,7 +31,8 @@ export const getElementUpdates = (
 
   const updatedProp = Object.getOwnPropertyDescriptor(Object.getPrototypeOf(element), propertyKey) as any;
 
-  if (updatedProp) {
+  //  Jest and JSDom breaks defining a new property, so skip
+  if (updatedProp && !isJestTest()) {
     Object.defineProperty(element, propertyKey, {
       get: updatedProp.get,
       set: val => {
