+++ packages/core/src/internal/utils/exists.spec.ts
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2016-2020 VMware, Inc. All Rights Reserved.
+ * Copyright (c) 2016-2022 VMware, Inc. All Rights Reserved.
  * This software is released under MIT license.
  * The full license information can be found in LICENSE in the root directory of this project.
  */
 
-import { existsIn, existsInWindow, isBrowser } from './exists.js';
+import { existsIn, existsInWindow } from './exists.js';
 
 describe('Functional Helper: ', () => {
   describe('existsInWindow(): ', () => {
@@ -89,12 +89,6 @@ describe('Functional Helper: ', () => {
 
     it('returns false shallowly', () => {
       expect(existsIn(['notDefined'], myTestObject)).toEqual(false);
-    });
-  });
-
-  describe('isBrowser():', () => {
-    it('returns true when expected', () => {
-      expect(isBrowser()).toBe(true);
     });
   });
 });
