+++ packages/core/src/internal/utils/exists.ts
@@ -1,11 +1,10 @@
 /*
- * Copyright (c) 2016-2021 VMware, Inc. All Rights Reserved.
+ * Copyright (c) 2016-2022 VMware, Inc. All Rights Reserved.
  * This software is released under MIT license.
  * The full license information can be found in LICENSE in the root directory of this project.
  */
 
 import curryN from 'ramda/es/curryN.js';
-import isNil from 'ramda/es/isNil.js';
 import path from 'ramda/es/path.js';
 import __ from './__.js';
 
@@ -26,7 +25,3 @@ export function elementExists(tagName: string, registry?: any): boolean {
 }
 
 export const existsInWindow = existsIn(__, window);
-
-export function isBrowser(win = window) {
-  return !isNil(win);
-}
