+++ packages/core/src/internal/utils/registration.ts
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2016-2021 VMware, Inc. All Rights Reserved.
+ * Copyright (c) 2016-2022 VMware, Inc. All Rights Reserved.
  * This software is released under MIT license.
  * The full license information can be found in LICENSE in the root directory of this project.
  */
 
 import curryN from 'ramda/es/curryN.js';
-import { elementExists, existsInWindow, isBrowser } from './exists.js';
+import { isBrowser } from './environment.js';
+import { elementExists, existsInWindow } from './exists.js';
 import { CDSState, setupCDSGlobal } from './global.js';
 import { isStorybook } from './framework.js';
 import { LogService } from '../services/log.service.js';
